+++ src/background/tabs-messaging.ts
@@ -160,6 +160,26 @@ browser.runtime.onMessage.addListener(
 				return { tabId };
 			}
 
+			case "clickHintInFrame":
+				await sendRequestToActiveTab({
+					type: "clickElement",
+					target: request.hint,
+				});
+				break;
+
+			case "markHintsAsKeyboardReachable":
+				await browser.tabs.sendMessage(tabId, {
+					type: "markHintsAsKeyboardReachable",
+					letter: request.letter,
+				});
+				break;
+
+			case "restoreKeyboardReachableHints":
+				await browser.tabs.sendMessage(tabId, {
+					type: "restoreKeyboardReachableHints",
+				});
+				break;
+
 			default:
 				throw new Error("Bad request to background script");
 		}