+++ src/content/actions/click-element.ts
@@ -1,10 +1,12 @@
 import { parseDomain, ParseResultType, fromUrl } from "parse-domain";
-import { isHintedIntersector } from "../../typing/typing-utils";
 import { focusesOnclick } from "../utils/clickable-type";
 import { flashHint } from "../hints/styles";
-import { getIntersectorWithHint } from "../intersectors";
+import {
+	getIntersectorsWithHints,
+	getIntersectorWithHint,
+} from "../intersectors";
 import { triggerHintsUpdate } from "../hints/display-hints";
-import { openInNewTab } from "./open-in-new-tab";
+import { openInBackgroundTab, openInNewTab } from "./open-in-new-tab";
 
 function getMainDomain(url: string): string | undefined {
 	const parseResult = parseDomain(fromUrl(url));
@@ -20,45 +22,71 @@ function getMainDomain(url: string): string | undefined {
 	return undefined;
 }
 
-export async function clickElement(hintText: string) {
-	const intersector = getIntersectorWithHint(hintText);
+function dispatchClick(element: Element) {
+	const mousedownEvent = new MouseEvent("mousedown", {
+		view: window,
+		bubbles: true,
+		cancelable: true,
+	});
+	const mouseupEvent = new MouseEvent("mouseup", {
+		view: window,
+		bubbles: true,
+		cancelable: true,
+	});
+	const clickEvent = new MouseEvent("click", {
+		view: window,
+		bubbles: true,
+		cancelable: true,
+	});
+	element.dispatchEvent(mousedownEvent);
+	element.dispatchEvent(mouseupEvent);
+	element.dispatchEvent(clickEvent);
+}
 
-	if (isHintedIntersector(intersector)) {
+export async function clickElement(hintTexts: string | string[]) {
+	let intersectors =
+		typeof hintTexts === "string"
+			? [getIntersectorWithHint(hintTexts)]
+			: getIntersectorsWithHints(hintTexts);
+
+	// If there are multiple targets and some of them are anchor elements we open
+	// those in a new background tab
+	if (intersectors.length > 1) {
+		const anchorIntersectorsHints = intersectors
+			.filter((intersector) => intersector.element instanceof HTMLAnchorElement)
+			.map((intersector) => intersector.hintText);
+		intersectors = intersectors.filter(
+			(intersector) => !(intersector.element instanceof HTMLAnchorElement)
+		);
+		await openInBackgroundTab(anchorIntersectorsHints);
+	}
+
+	for (const intersector of intersectors) {
 		const element = intersector.element;
 		flashHint(intersector);
-		if (focusesOnclick(element)) {
-			(
-				element as HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement
-			).focus();
-		} else {
-			if (element instanceof HTMLAnchorElement) {
-				const linkElement = element;
-				// Sometimes websites use links with target="_blank" but don't open a new tab.
-				// They probably prevent the default behavior with javascript. For example Slack
-				// has this for opening a thread in the side panel. So here we make sure that
-				// if the main domains are equal just do a normal click and let the page handle it
-				const linkMainDomain = getMainDomain(linkElement.href);
-				const locationMainDomain = getMainDomain(window.location.href);
-				if (
-					linkElement.getAttribute("target") === "_blank" &&
-					linkMainDomain !== locationMainDomain
-				) {
-					await openInNewTab(intersector);
-				} else {
-					linkElement.click();
-				}
+		if (element instanceof HTMLElement && focusesOnclick(element)) {
+			element.focus();
+		} else if (element instanceof HTMLAnchorElement) {
+			// Sometimes websites use links with target="_blank" but don't open a new tab.
+			// They probably prevent the default behavior with javascript. For example Slack
+			// has this for opening a thread in the side panel. So here we make sure that
+			// if the main domains are equal just do a normal click and let the page handle it
+			const linkMainDomain = getMainDomain(element.href);
+			const locationMainDomain = getMainDomain(window.location.href);
+			if (
+				element.getAttribute("target") === "_blank" &&
+				linkMainDomain !== locationMainDomain
+			) {
+				void openInNewTab(intersector);
 			} else {
-				const event = new MouseEvent("click", {
-					view: window,
-					bubbles: true,
-					cancelable: true,
-				});
-				element.dispatchEvent(event);
+				dispatchClick(element);
 			}
-
-			// On some pages like codepen there are hints remaining after closing a popup panel.
-			// This is not a perfect solution but as long as the user clicks with voice I think we're safe
-			await triggerHintsUpdate();
+		} else {
+			dispatchClick(element);
 		}
 	}
+
+	// On some pages like codepen there are hints remaining after closing a popup panel.
+	// This is not a perfect solution but as long as the user clicks with voice I think we're safe
+	await triggerHintsUpdate();
 }