+++ src/content/actions/scroll.ts
@@ -1,75 +1,170 @@
 import { getIntersectorWithHint } from "../intersectors";
+import { getScrollContainer } from "../utils/get-scroll-container";
 
-let scrollContainer: Element | undefined;
+let scrollContainer: HTMLElement | undefined;
+let lastScrollFactor: number;
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-function getScrollContainer(node: Node | null): Element | undefined {
-	if (node instanceof Element) {
-		if (
-			node.scrollHeight > node.clientHeight &&
-			(window.getComputedStyle(node).overflowY === "auto" ||
-				window.getComputedStyle(node).overflowY === "scroll")
-		) {
-			if (
-				node === document.body &&
-				document.documentElement.clientHeight !==
-					document.documentElement.scrollHeight
-			) {
-				return document.documentElement;
-			}
-
-			return node;
-		}
-
-		if (node.parentNode) {
-			return getScrollContainer(node.parentNode);
-		}
-	}
-
-	return undefined;
+function isPageScroll(container: Element) {
+	return container === document.body || container === document.documentElement;
 }
 
-function scrollVertically(
+function scrollVerticallyAmount(
 	scrollContainer: HTMLElement,
-	direction: "up" | "down"
+	scrollAmount: number
 ) {
 	const previousScrollBehavior =
 		window.getComputedStyle(scrollContainer).scrollBehavior;
 	scrollContainer.style.scrollBehavior = "auto";
+
+	scrollContainer.scrollBy({ left: 0, top: scrollAmount, behavior: "auto" });
+	scrollContainer.style.scrollBehavior = previousScrollBehavior;
+}
+
+function scrollVertically(
+	scrollContainer: HTMLElement,
+	direction: "up" | "down",
+	scrollFactor: number
+) {
+	lastScrollFactor = scrollFactor;
 	const scrollHeight = Math.min(
 		scrollContainer.clientHeight,
 		document.documentElement.clientHeight
 	);
 	const scrollAmount =
-		direction === "up" ? (-scrollHeight * 2) / 3 : (scrollHeight * 2) / 3;
-	scrollContainer.scrollBy({ left: 0, top: scrollAmount, behavior: "auto" });
-	scrollContainer.style.scrollBehavior = previousScrollBehavior;
+		direction === "up"
+			? -scrollHeight * scrollFactor
+			: scrollHeight * scrollFactor;
+
+	scrollVerticallyAmount(scrollContainer, scrollAmount);
 }
 
 export function scrollVerticallyAtElement(
 	direction: "up" | "down",
-	hint?: string
+	hint?: string,
+	scrollFactor?: number
 ) {
+	scrollFactor = scrollFactor ?? lastScrollFactor ?? 0.66;
 	if (hint) {
 		const element = getIntersectorWithHint(hint).element;
 		scrollContainer = getScrollContainer(element);
 	}
 
-	if (!scrollContainer) {
-		scrollPageVertically(direction);
-	}
-
-	if (scrollContainer && scrollContainer instanceof HTMLElement) {
-		scrollVertically(scrollContainer, direction);
+	if (scrollContainer) {
+		scrollVertically(scrollContainer, direction, scrollFactor);
 	}
 }
 
-export function scrollPageVertically(direction: "up" | "down") {
+function getElementVisibleRect(element: Element): DOMRect {
+	const rect = element.getBoundingClientRect();
+	const viewportHeight = document.documentElement.clientHeight;
+	const viewportWidth = document.documentElement.clientWidth;
+	const elementTop = Math.max(0, rect.top);
+	const elementBottom = Math.min(viewportHeight, rect.bottom);
+	const elementLeft = Math.max(0, rect.left);
+	const elementRight = Math.min(0, viewportWidth);
+	return new DOMRect(
+		elementLeft,
+		elementTop,
+		elementRight - elementLeft,
+		elementBottom - elementTop
+	);
+}
+
+export function scrollElementToTop(hint: string) {
+	const element = getIntersectorWithHint(hint).element;
+	scrollContainer = getScrollContainer(element);
+
+	if (scrollContainer) {
+		const containerRect = getElementVisibleRect(scrollContainer);
+		const elementTop = element.getBoundingClientRect().top;
+		const scrollAmount = isPageScroll(scrollContainer)
+			? elementTop
+			: elementTop - containerRect.top;
+
+		scrollVerticallyAmount(scrollContainer, scrollAmount);
+
+		// After scrolling we need to check if were the element sits now there is a
+		// sticky or fixed element obscuring the element. If that's the case we scroll
+		// down the height of that sticky element
+		const elementRect = element.getBoundingClientRect();
+		const elementsAt = document.elementsFromPoint(
+			elementRect.x + 5,
+			elementRect.y + 5
+		);
+		let outerSticky;
+
+		for (const elementAt of elementsAt) {
+			if (elementAt === element || elementAt.contains(element)) {
+				break;
+			}
+
+			const elementPosition = window.getComputedStyle(elementAt).position;
+			if (
+				(elementPosition === "sticky" || elementPosition === "fixed") &&
+				(!outerSticky ||
+					outerSticky.compareDocumentPosition(elementAt) === 4 || // ElementAt precedes outerSticky
+					outerSticky.compareDocumentPosition(elementAt) === 10) // ElementAt precedes and contains outerSticky
+			) {
+				outerSticky = elementAt;
+			}
+		}
+
+		if (outerSticky) {
+			scrollVerticallyAmount(scrollContainer, -outerSticky.clientHeight);
+		}
+	}
+}
+
+export function scrollElementToBottom(hint: string) {
+	const element = getIntersectorWithHint(hint).element;
+	scrollContainer = getScrollContainer(element);
+
+	if (scrollContainer) {
+		const containerRect = getElementVisibleRect(scrollContainer);
+		const elementBottom = element.getBoundingClientRect().bottom;
+		const scrollAmount = isPageScroll(scrollContainer)
+			? scrollContainer.clientHeight - elementBottom
+			: containerRect.bottom - elementBottom;
+
+		scrollVerticallyAmount(scrollContainer, -scrollAmount);
+	}
+}
+
+export function scrollElementToCenter(hint: string) {
+	const element = getIntersectorWithHint(hint).element;
+	scrollContainer = getScrollContainer(element);
+
+	if (scrollContainer) {
+		const containerRect = getElementVisibleRect(scrollContainer);
+		const containerCenter = isPageScroll(scrollContainer)
+			? scrollContainer.clientHeight / 2
+			: containerRect.top + containerRect.height / 2;
+		const elementCenter =
+			element.getBoundingClientRect().top + element.clientHeight / 2;
+		const scrollAmount =
+			Math.max(containerCenter, elementCenter) -
+			Math.min(containerCenter, elementCenter);
+
+		// If the container center is greater than the element center, that means that
+		// the element is below the center of the container. In that case the scroll
+		// amount needs to be negative to move the element upwards.
+		if (containerCenter > elementCenter) {
+			scrollVerticallyAmount(scrollContainer, -scrollAmount);
+		} else {
+			scrollVerticallyAmount(scrollContainer, scrollAmount);
+		}
+	}
+}
+
+export function scrollPageVertically(
+	direction: "up" | "down",
+	scrollAmount = 0.66
+) {
 	const scrollContainer =
 		document.documentElement.clientHeight ===
 		document.documentElement.scrollHeight
 			? document.body
 			: document.documentElement;
 
-	scrollVertically(scrollContainer, direction);
+	scrollVertically(scrollContainer, direction, scrollAmount);
 }