+++ src/content/actions/show.ts
@@ -1,10 +1,35 @@
 import { getIntersectorWithHint } from "../intersectors";
 import { showTooltip } from "../hints/tooltip";
+import { flashHint } from "../hints/styles";
 
-export function showLink(hintText: string) {
+export function showTitleAndHref(hintText: string) {
 	const target = getIntersectorWithHint(hintText);
-	if (target) {
-		const href = (target.element as HTMLLinkElement).href;
-		showTooltip(target, href, 5000);
+	const element = target.element;
+	let result = "";
+	let title = "";
+
+	if (element instanceof HTMLElement) {
+		if (element.title) {
+			title = element.title;
+		} else {
+			const subnodeWithTitle = element.querySelector("[title");
+			if (subnodeWithTitle instanceof HTMLElement && subnodeWithTitle.title) {
+				title = subnodeWithTitle.title;
+			}
+		}
+	}
+
+	if (element instanceof HTMLAnchorElement) {
+		result = title
+			? `<div><strong>${title}</strong></div><div>${element.href}</div>`
+			: `<div>${element.href}</div>`;
+	} else {
+		result = title;
+	}
+
+	if (result) {
+		showTooltip(target, result, 5000);
+	} else {
+		flashHint(target);
 	}
 }