+++ src/content/hints/position-hints.ts
@@ -54,7 +54,7 @@ export function positionHint(intersector: HintedIntersector) {
 		document.documentElement.scrollTop ||
 		document.body.scrollTop;
 
-	const nudgeX = 0.3;
+	const nudgeX = intersector.hintAnchorIsText ? 0.1 : 0.3;
 	const nudgeY = 0.4;
 
 	let x = rect.left + scrollLeft - hintElement.offsetWidth * (1 - nudgeX);
@@ -94,4 +94,6 @@ export function positionHint(intersector: HintedIntersector) {
 			hintElement.style.top = `${scrollTop + targetY}px`;
 		}
 	}
+
+	intersector.hintElement.style.display = "block";
 }